added content list item view

+++ app/src/main/java/weather/co/detail/model/Wind.java
@@ -13,7 +13,7 @@ public class Wind {
     private float speed;
     @SerializedName("deg")
     @Expose
-    private int deg;
+    private float deg;
 
     public float getSpeed() {
         return speed;
@@ -23,11 +23,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(float deg) {
         this.deg = deg;
     }
